Refactored ATM and House Functionality
